Updated rofi

+++ dotfiles-manager.sh
@@ -1,13 +1,25 @@
 #bin/bash
 #qtile
-rm -rf ~/Utilities/git-cloning/Dotfiles_V2/qtile/
-cp -r ~/.config/qtile ~/Utilities/git-cloning/Dotfiles_V2/
+if [ $1 == 'qtile' ]
+then
+	rm -rf ~/Utilities/git-cloning/Dotfiles_V2/qtile/
+	cp -r ~/.config/qtile ~/Utilities/git-cloning/Dotfiles_V2/
+fi
 #neovim
-rm -rf ~/Utilities/git-cloning/Dotfiles_V2/neovim/
-cp -r ~/.config/nvim ~/Utilities/git-cloning/Dotfiles_V2/
+if [ $1 == 'nvim' ]
+then
+	rm -rf ~/Utilities/git-cloning/Dotfiles_V2/neovim/
+	cp -r ~/.config/nvim ~/Utilities/git-cloning/Dotfiles_V2/
+fi
 #rofi
-rm -rf ~/Utilities/git-cloning/Dotfiles_V2/rofi/
-cp -r ~/.config/rofi/ ~/Utilities/git-cloning/Dotfiles_V2/
+if [ $1 == 'rofi' ]
+then
+	rm -rf ~/Utilities/git-cloning/Dotfiles_V2/rofi/
+	cp -r ~/.config/rofi/ ~/Utilities/git-cloning/Dotfiles_V2/
+fi
 #kitty
-rm -rf ~/Utilities/git-cloning/Dotfiles_V2/kitty/
-cp -r ~/.config/kitty/ ~/Utilities/git-cloning/Dotfiles_V2/
+if [ $1 == 'qtile' ]
+then
+	rm -rf ~/Utilities/git-cloning/Dotfiles_V2/kitty/
+	cp -r ~/.config/kitty/ ~/Utilities/git-cloning/Dotfiles_V2/
+fi
